Add, normalize and clean up data
The data separated stored in the two files aren't well normalized. For
example, the biomarkers stored in the LabValues file belong to each
patient.

Therefore, clean this up at load time into a cleaner interface.

Data checking happens in testthat files. This is because we have static
files known at development time. We get less complexity in the rest of
the app by assuming clean, good data. If we were are to switch the
storage into something more dynamic like a database or file upload,
we'll need to move this into the live code.